requirements.txt
# Web Framework
flask==2.3.3
werkzeug==2.3.7

# Database
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
sqlalchemy==2.0.21
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Authorization
flask-security-too==5.4.3
flask-login==0.6.3
flask-principal==0.4.0

# Forms & Validation
flask-wtf==1.2.1
wtforms==3.0.1
email-validator==2.1.0

# Security
flask-talisman==1.1.0
flask-limiter==3.5.0
bcrypt==4.1.2
argon2-cffi==23.1.0

# API
flask-restful==0.3.10
flask-jwt-extended==4.6.0
flask-cors==6.0.1
marshmallow==3.20.1
apispec==6.3.0

# Email
flask-mail==0.9.1

# Environment & Configuration
python-dotenv==1.0.0

# Logging
structlog==23.2.0

# Machine Learning (existing)
scikit-learn==1.3.2
numpy==1.26.0
pandas==2.1.1
joblib==1.3.2

# Testing
pytest==7.4.2
pytest-flask==1.3.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Development Tools
black==23.11.0
isort==5.13.0
flake8==6.1.0
mypy==1.7.1

# Deployment
gunicorn==21.2.0
redis==5.0.1

# Monitoring
sentry-sdk[flask]==1.38.0